PR feedbacks
Moved UIWindow extension to an independant file, updated podspec to take objective C files refactoring in account
