+++ Sources/XLPagerTabStrip/TwitterPagerTabStripViewController.swift
@@ -28,31 +28,6 @@ import UIKit
 import FXPageControl
 #endif
 
-extension UIWindow {
-    static var isLandscape: Bool {
-        if #available(iOS 13.0, *) {
-            return UIApplication.shared.windows
-                .first?
-                .windowScene?
-                .interfaceOrientation
-                .isLandscape ?? false
-        } else {
-            return UIApplication.shared.statusBarOrientation.isLandscape
-        }
-    }
-    static var isPortrait: Bool {
-        if #available(iOS 13.0, *) {
-            return UIApplication.shared.windows
-                .first?
-                .windowScene?
-                .interfaceOrientation
-                .isPortrait ?? false
-        } else {
-            return UIApplication.shared.statusBarOrientation.isPortrait
-        }
-    }
-}
-
 public struct TwitterPagerTabStripSettings {
 
     public struct Style {
